+++ rubato/utils/draw.py
@@ -1,7 +1,7 @@
 """A static draw class for drawing things directly to the renderer."""
 from __future__ import annotations
 from ctypes import c_int16
-from typing import List, Optional, Callable
+from typing import TYPE_CHECKING, List, Optional, Callable
 from dataclasses import dataclass, field
 import heapq
 
@@ -9,6 +9,9 @@ import sdl2, sdl2.ext
 from sdl2.sdlgfx import pixelRGBA, thickLineRGBA, filledPolygonRGBA, aapolygonRGBA
 
 from . import Vector, Color, Font, Display, Math
+
+if TYPE_CHECKING:
+    from . import Sprite
 
 
 @dataclass(order=True)
@@ -359,24 +362,47 @@ class Draw:
         Display.update(tx, pos + (align - 1) * Vector(*tx.size) / 2)
 
     @classmethod
-    def texture(cls, image: sdl2.ext.Texture, pos: Vector = Vector(), z_index: int = Math.INF):
+    def texture(cls, texture: sdl2.ext.Texture, pos: Vector = Vector(), z_index: int = Math.INF):
         """
-        Draws an image onto the renderer at the end of the frame.
+        Draws an texture onto the renderer at the end of the frame.
 
         Args:
-            image: The image to draw.
-            pos: The position of the image. Defaults to Vector(0, 0).
+            texture: The texture to draw.
+            pos: The position of the texture. Defaults to Vector(0, 0).
             z_index: Where to draw it in the drawing order. Defaults to Math.INF.
         """
-        heapq.heappush(cls._queue, DrawTask(z_index, lambda: cls.immediate_texture(image, pos)))
+        heapq.heappush(cls._queue, DrawTask(z_index, lambda: cls.immediate_texture(texture, pos)))
 
     @staticmethod
-    def immediate_texture(image: sdl2.ext.Texture, pos: Vector = Vector()):
+    def immediate_texture(texture: sdl2.ext.Texture, pos: Vector = Vector()):
         """
         Draws an SDL Texture onto the renderer immediately.
 
         Args:
-            image: The texture to draw.
+            texture: The texture to draw.
             pos: The position to draw the texture at. Defaults to Vector().
         """
-        Display.update(image, pos)
+        Display.update(texture, pos)
+
+    @classmethod
+    def sprite(cls, sprite: Sprite, pos: Vector = Vector(), z_index: int = 0):
+        """
+        Draws an sprite onto the renderer at the end of the frame.
+
+        Args:
+            sprite: The sprite to draw.
+            pos: The position to draw the sprite at. Defaults to Vector(0, 0).
+            z_index: The z-index of the sprite. Defaults to 0.
+        """
+        heapq.heappush(cls._queue, DrawTask(z_index, lambda: cls.immediate_sprite(sprite, pos)))
+
+    @staticmethod
+    def immediate_sprite(sprite: Sprite, pos: Vector = Vector()):
+        """
+        Draws an sprite onto the renderer immediately.
+
+        Args:
+            sprite: The sprite to draw.
+            pos: The position to draw the sprite at. Defaults to Vector().
+        """
+        sprite.render(pos)